requirements.txt
streamlit
langchain
langchain-community 
langchain-experimental 
chromadb
transformers
streamlit_chat
langchain_groq
langchain_cohere
langchain_core
tiktoken 
python-dotenv
ollama
chromadb
streamlit_chat
numpy<2
pypika==0.48.0



